# projects/ecommerce_store/backend/backend_app/api/routes/order.py
# GENERATED by the Backend Agent from the project graphs. Do not edit by hand: regenerate instead.
"""HTTP routes for service.order. Route handlers only delegate to the application service."""
from __future__ import annotations

from typing import Literal
from uuid import UUID

from fastapi import APIRouter, Depends, Query, Response

from ...api_common import error_responses
from ...auth import authorize
from ...engine import ListQuery, Principal
from ..deps import get_services
from ...application.services import Services
from ..schemas.order import OrderPlaceRequest, OrderResponse


router = APIRouter(tags=["order"])


@router.get(
    "/orders",
    operation_id="api.order.list",
    summary="List Orders",
    description="List Orders.\n\nOperation `operation.order.list`. Access: restricted (customer).",
    response_model=list[OrderResponse],
    status_code=200,
    responses=error_responses(authenticated=True, restricted=True, has_path_id=False, mutating=False, rate_limited=False),
)
def api_order_list(
    response: Response,
    limit: int = Query(100, ge=1, le=1000),
    offset: int = Query(0, ge=0),
    sort: str | None = Query(None, max_length=64, description="Attribute to sort by"),
    order: Literal['asc', 'desc'] = Query('asc'),
    principal: Principal | None = Depends(authorize("operation.order.list")),
    services: Services = Depends(get_services),
):
    page = services.order.list(principal, ListQuery(filters={}, search=None, sort=sort, order=order, limit=limit, offset=offset))
    response.headers['X-Total-Count'] = str(page.total)
    return page.items


@router.post(
    "/orders",
    operation_id="api.order.place",
    summary="Place order",
    description="Place order.\n\nOperation `operation.order.place`. Access: restricted (customer). Note: required attributes with no source: ['total']; takes non-attribute inputs ['shipping_address', 'payment_method']",
    response_model=OrderResponse,
    status_code=201,
    responses=error_responses(authenticated=True, restricted=True, has_path_id=False, mutating=True, rate_limited=False),
)
def api_order_place(
    body: OrderPlaceRequest,
    principal: Principal | None = Depends(authorize("operation.order.place")),
    services: Services = Depends(get_services),
):
    return services.order.place(principal, body=body.model_dump(exclude_unset=True))


@router.get(
    "/orders/{id}",
    operation_id="api.order.read",
    summary="View Order",
    description="View Order.\n\nOperation `operation.order.read`. Access: restricted (customer).",
    response_model=OrderResponse,
    status_code=200,
    responses=error_responses(authenticated=True, restricted=True, has_path_id=True, mutating=False, rate_limited=False),
)
def api_order_read(
    id: UUID,
    principal: Principal | None = Depends(authorize("operation.order.read")),
    services: Services = Depends(get_services),
):
    return services.order.read(principal, id)