# projects/ecommerce_store/backend/backend_app/api/schemas/order.py
# GENERATED by the Backend Agent from the project graphs. Do not edit by hand: regenerate instead.
"""API request/response models for entity.order (from api.json schemas)."""
from __future__ import annotations

from datetime import datetime
from typing import Literal
from uuid import UUID

from pydantic import BaseModel, ConfigDict, Field

from ...api_common import JsonDecimal


class OrderResponse(BaseModel):
    """schema.order: Order"""

    model_config = ConfigDict(from_attributes=True)
    id: UUID
    user_id: UUID
    status: Literal['pending', 'paid', 'shipped', 'delivered', 'cancelled']
    total: JsonDecimal
    shipping_address: str | None = None
    placed_at: datetime | None = None


class OrderPlaceRequest(BaseModel):
    """schema.order.place: Place order request"""

    model_config = ConfigDict(extra="forbid", str_strip_whitespace=True)
    shipping_address: str = Field(..., max_length=20000, min_length=1)
    payment_method: str = Field(..., max_length=255, min_length=1)